CORDA-3715: Now check class version when contract verification takes place.
