+++ node/src/main/kotlin/net/corda/node/internal/cordapp/JarScanningCordappLoader.kt
@@ -35,7 +35,6 @@ import java.util.jar.Manifest
 import java.util.zip.ZipInputStream
 import kotlin.reflect.KClass
 import kotlin.streams.toList
-import java.io.DataInputStream
 
 /**
  * Handles CorDapp loading and classpath scanning of CorDapp JARs
@@ -62,8 +61,6 @@ class JarScanningCordappLoader private constructor(private val cordappJarPaths: 
 
     companion object {
         private val logger = contextLogger()
-        private const val MIN_CONTRACT_CLASS_VERSION = 49 // Java 5
-        private const val MAX_CONTRACT_CLASS_VERSION = 52 // Java 8
 
         /**
          * Creates a CordappLoader from multiple directories.
@@ -189,31 +186,15 @@ class JarScanningCordappLoader private constructor(private val cordappJarPaths: 
 
     private fun CordappImpl.verifyClassVersion(): CordappImpl {
         if (contractClassNames.isNotEmpty()) {
-            verifyClassVersionNumber(contractClassNames, jarPath)
+            verifyClassVersionNumber(appClassLoader, contractClassNames, jarPath)
         }
         if (contractStateClassNames.isNotEmpty()) {
-            verifyClassVersionNumber(contractStateClassNames, jarPath)
+            verifyClassVersionNumber(appClassLoader, contractStateClassNames, jarPath)
         }
         if (serializationCustomSerializers.isNotEmpty()) {
-            verifyClassVersionNumber(serializationCustomSerializers.map { it.javaClass.name }.toList(), jarPath)
+            verifyClassVersionNumber(appClassLoader, serializationCustomSerializers.map { it.javaClass.name }.toList(), jarPath)
         }
         return this
-    }
-
-    private fun verifyClassVersionNumber(classes: List<String>, jarPath: URL) {
-        classes.forEach {
-            DataInputStream(appClassLoader.getResourceAsStream("${it.replace('.', '/')}.class")).use { dataStream ->
-                val magic = dataStream.readInt()
-                if (magic != -0x35014542) {
-                    throw java.lang.IllegalStateException("Invalid Java class $it found in jar file $jarPath")
-                }
-                dataStream.readShort() // minor version number
-                val major = dataStream.readShort().toInt()
-                if (major < MIN_CONTRACT_CLASS_VERSION || major > MAX_CONTRACT_CLASS_VERSION) {
-                    throw IllegalStateException("Class $it from jar file $jarPath has an invalid version of $major")
-                }
-            }
-        }
     }
 
     private fun parseCordappInfo(manifest: Manifest?, defaultName: String): Cordapp.Info {
